+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Add `capitalizeEachWords` method
 - Add `strCount` method
 - Add `sleep` method
+- Add `arrShuffle` method
 
 ## 0.0.1 (2024-09-02) - Not for Production